GCP cloud function-source/requirements.txt
functions-framework==3.*
requests
pandas
google-cloud-bigquery==3.2.0
google-cloud-secret-manager==2.11.0
google-cloud-storage==2.14.0
Python-IO
click==8.1.7
google-cloud-appengine-logging==1.3.2
    # via google-cloud-logging
google-cloud-audit-log==0.2.5
    # via google-cloud-logging
google-cloud-core==2.3.3
google-cloud-logging==3.8.0
    # via -r src/requirements.in